refactor: move custom functions for multisig creaiton from bdk into bip32/bip39
