+++ apps/mobile/app/(authenticated)/(tabs)/(signer,explorer,converter)/account/add/(common)/generate/mnemonic/[index].tsx
@@ -4,11 +4,7 @@ import { useState } from 'react'
 import { ScrollView } from 'react-native'
 import { useShallow } from 'zustand/react/shallow'
 
-import {
-  getDescriptorObject,
-  getExtendedPublicKeyFromMnemonic,
-  parseDescriptor
-} from '@/api/bdk'
+import { getDescriptorObject, parseDescriptor } from '@/api/bdk'
 import SSButton from '@/components/SSButton'
 import SSChecksumStatus from '@/components/SSChecksumStatus'
 import SSFingerprint from '@/components/SSFingerprint'
@@ -25,7 +21,11 @@ import { useAccountBuilderStore } from '@/store/accountBuilder'
 import { useBlockchainStore } from '@/store/blockchain'
 import { type GenerateMnemonicSearchParams } from '@/types/navigation/searchParams'
 import { getDescriptorsFromKey } from '@/utils/bip32'
-import { getFingerprintFromMnemonic, validateMnemonic } from '@/utils/bip39'
+import {
+  getExtendedPublicKeyFromMnemonicCustom,
+  getFingerprintFromMnemonic,
+  validateMnemonic
+} from '@/utils/bip39'
 import {
   getDerivationPathFromScriptVersion,
   getMultisigDerivationPathFromScriptVersion
@@ -109,7 +109,7 @@ export default function GenerateMnemonic() {
         )
         derivationPath = `m/${rawDerivationPath}`
         // Generate extended public key first using the same method as import flow
-        const extendedPublicKey = await getExtendedPublicKeyFromMnemonic(
+        const extendedPublicKey = await getExtendedPublicKeyFromMnemonicCustom(
           mnemonic.join(' '),
           passphrase || '',
           network as Network,
