+++ apps/mobile/app/(authenticated)/(tabs)/(signer,explorer,converter)/account/add/(common)/import/mnemonic/[keyIndex].tsx
@@ -5,7 +5,6 @@ import { ScrollView } from 'react-native'
 import { toast } from 'sonner-native'
 import { useShallow } from 'zustand/react/shallow'
 
-import { getExtendedPublicKeyFromMnemonic } from '@/api/bdk'
 import SSEllipsisAnimation from '@/components/SSEllipsisAnimation'
 import SSGradientModal from '@/components/SSGradientModal'
 import SSSeedWordsInput from '@/components/SSSeedWordsInput'
@@ -23,6 +22,7 @@ import { useBlockchainStore } from '@/store/blockchain'
 import { Colors } from '@/styles'
 import { type Account } from '@/types/models/Account'
 import { type ImportMnemonicSearchParams } from '@/types/navigation/searchParams'
+import { getExtendedPublicKeyFromMnemonicCustom } from '@/utils/bip39'
 import { getScriptVersionDisplayName } from '@/utils/scripts'
 export default function ImportMnemonic() {
   const { keyIndex } = useLocalSearchParams<ImportMnemonicSearchParams>()
@@ -147,7 +147,7 @@ export default function ImportMnemonic() {
       // For multisig, we need to generate the extended public key from the mnemonic
       if (currentMnemonic && currentFingerprint) {
         // Generate the extended public key
-        const extendedPublicKey = await getExtendedPublicKeyFromMnemonic(
+        const extendedPublicKey = await getExtendedPublicKeyFromMnemonicCustom(
           currentMnemonic,
           passphrase || '',
           network as Network,
